ScheduledJob update + test update

+++ invoice-app/src/main/java/bg/softuni/invoice_app/utils/ScheduledJob.java
@@ -1,6 +1,8 @@
 package bg.softuni.invoice_app.utils;
 
+import bg.softuni.invoice_app.model.entity.ArchiveInvoice;
 import bg.softuni.invoice_app.model.entity.BankAccountPersist;
+import bg.softuni.invoice_app.service.archive.ArchiveInvoiceService;
 import bg.softuni.invoice_app.service.bankAccountPersist.BankAccountPersistService;
 import bg.softuni.invoice_app.service.invoice.InvoiceService;
 import jakarta.transaction.Transactional;
@@ -14,10 +16,12 @@ public class ScheduledJob {
   
   private final BankAccountPersistService bankAccountPersistService;
   private final InvoiceService invoiceService;
+  private final ArchiveInvoiceService archiveInvoiceService;
   
-  public ScheduledJob(BankAccountPersistService bankAccountPersistService, InvoiceService invoiceService) {
+  public ScheduledJob(BankAccountPersistService bankAccountPersistService, InvoiceService invoiceService, ArchiveInvoiceService archiveInvoiceService) {
     this.bankAccountPersistService = bankAccountPersistService;
     this.invoiceService = invoiceService;
+    this.archiveInvoiceService = archiveInvoiceService;
   }
   
   @Transactional
@@ -27,7 +31,8 @@ public class ScheduledJob {
     
     for (BankAccountPersist account : allPersistantAccounts) {
       boolean isUsed = invoiceService.existsByBankAccount(account);
-      if (!isUsed) {
+      boolean isUsedInArchive = archiveInvoiceService.existsByBankAccount(account);
+      if (!isUsed && !isUsedInArchive) {
         bankAccountPersistService.delete(account);
       }
     }
